Preliminary integration of Image into ImageView.
