+++ ui/src/main/scala/io/youi/image/SVGImage.scala
@@ -1,0 +1,56 @@
+package io.youi.image
+
+import com.outr.CanvgImplicits._
+import io.youi.dom._
+import io.youi.component.Component
+import io.youi.component.draw.BoundingBox
+import io.youi.component.draw.path.Path
+import org.scalajs.dom.raw._
+
+case class SVGImage(svg: SVGSVGElement,
+                    width: Double,
+                    height: Double,
+                    original: Option[Image]) extends Image {
+  override def drawImage(component: Component, context: CanvasRenderingContext2D, width: Double, height: Double): Unit = {
+    context.drawSvg(svg.outerHTML, 0.0, 0.0, width, height)
+  }
+}
+
+object SVGImage {
+  def measure(svg: SVGElement): BoundingBox = {
+    var minX = 0.0
+    var minY = 0.0
+    var maxX = 0.0
+    var maxY = 0.0
+    def measureInternal(e: Element, offsetX: Double, offsetY: Double): Unit = e match {
+      case g: SVGGElement => {
+        var ox = offsetX
+        var oy = offsetY
+        (0 until g.transform.baseVal.numberOfItems).foreach { index =>
+          val transform = g.transform.baseVal.getItem(index)
+          if (transform.`type` == SVGTransform.SVG_TRANSFORM_TRANSLATE) {
+            ox += transform.matrix.e
+            oy += transform.matrix.f
+          }
+        }
+        g.children.foreach(child => measureInternal(child, ox, oy))
+      }
+      case c: SVGCircleElement => {
+        minX = math.min(minX, offsetX + (c.cx.baseVal.value - c.r.baseVal.value))
+        minY = math.min(minY, offsetY + (c.cy.baseVal.value - c.r.baseVal.value))
+        maxX = math.max(maxX, offsetX + (c.cx.baseVal.value + c.r.baseVal.value))
+        maxY = math.max(maxY, offsetY + (c.cy.baseVal.value + c.r.baseVal.value))
+      }
+      case p: SVGPathElement => {
+        val path = Path(p.getAttribute("d"))
+        minX = math.min(minX, offsetX + path.boundingBox.x1)
+        minY = math.min(minY, offsetY + path.boundingBox.y1)
+        maxX = math.max(maxX, offsetX + path.boundingBox.x2)
+        maxY = math.max(maxY, offsetY + path.boundingBox.y2)
+      }
+      case _ => scribe.warn(s"Unsupported SVG node: $e.")
+    }
+    svg.children.foreach(child => measureInternal(child, 0.0, 0.0))
+    BoundingBox(minX, minY, maxX, maxY)
+  }
+}
